WEB3-22 - Some changes to design
Two latest commits has been for a new navigation bar which will be to the left, only available on mobile devices and activated by clicking on the menu-icon.

The footer will disappear on mobile devices and the search bar will appear in the sidebar menu instead of its original position.